refactor(forms): merge the in/out Dial, Multiplex and Listen forms
Completes the in/out consolidation. out/ is the surviving copy in all three:
it guards its inline switches with `?? false` where in/ bound the raw field
(which is what produced the "Expected Boolean, got Undefined" warnings), wraps
its popover entries in .pop-item for hover feedback, and pins the popover
direction.

Two behaviours had to move across:

  - out/Dial gains mode?: string. mode="client" hides the seven server-only
    toggles plus domain resolver, leaving TCP/UDP options, connect timeout and
    keep-alive — verified by opening the popover and counting: 12 entries
    without it, exactly those 4 with it, matching in/Dial's ungated set.
  - out/Multiplex gains direction?: string, gating protocol, max connections
    and the stream bounds — the client's half of the negotiation. Gated on
    `!= 'in'` rather than `== 'out'` so the nine outbound callers, none of
    which pass a direction, keep those fields. Counted: 4 fields with no
    direction, 4 with "out", 0 with "in".

Listen has no such switch and simply adopts out/'s layout, which is a visible
change in the inbound drawer: detour and UDP NAT expiration join the address
and port grid rather than sitting on their own rows, and TCP Fast Open, TCP
Multi Path and UDP Fragment share one switch row instead of two.

components/forms/in is now 15 files, from 26 before this pass.
