+++ frontend/src/layouts/drawers/inbound/InboundDrawer.vue
@@ -113,7 +113,7 @@ import Chip from '@/components/ui/Chip.vue'
 import Btn from '@/components/ui/Btn.vue'
 import Ico from '@/components/ui/Ico.vue'
 import IconBtn from '@/components/ui/IconBtn.vue'
-import Listen from '@/components/forms/in/Listen.vue'
+import Listen from '@/components/forms/out/Listen.vue'
 import Direct from '@/components/forms/in/Direct.vue'
 import Shadowsocks from '@/components/forms/in/Shadowsocks.vue'
 import Hysteria from '@/components/forms/in/Hysteria.vue'
@@ -127,9 +127,9 @@ import TProxy from '@/components/forms/in/TProxy.vue'
 import Transport from '@/components/forms/out/Transport.vue'
 import Users from '@/components/forms/in/Users.vue'
 import InTLS from '@/components/forms/in/InTLS.vue'
-import Multiplex from '@/components/forms/in/Multiplex.vue'
+import Multiplex from '@/components/forms/out/Multiplex.vue'
 import OutJson from '@/components/forms/in/OutJson.vue'
-import Dial from '@/components/forms/in/Dial.vue'
+import Dial from '@/components/forms/out/Dial.vue'
 import AddrVue from '@/components/forms/in/Addr.vue'
 
 const props = defineProps<{
